Separated out seeding test
